Interconnect waku and nomos nodes via the same method
